Move Artemii's streamer to op-node/rollup/derive and split batch validity check and actions to perform with the batch.
